minor update, sprite add typeCast to float, mainShip switch HeightPrortion and region

+++ engine/src/ru/geekbrains/stargame/engine/Sprite.java
@@ -35,13 +35,13 @@ public class Sprite extends Rect {
 
     public void setWidthProportion(float width) {
         setWidth(width);
-        float aspect = regions[frame].getRegionWidth() / regions[frame].getRegionHeight();
+        float aspect = regions[frame].getRegionWidth() / (float)regions[frame].getRegionHeight();
         setHeight(width / aspect);
     }
 
     public void setHeightProportion(float height) {
         setHeight(height);
-        float aspect = regions[frame].getRegionWidth() / regions[frame].getRegionHeight();
+        float aspect = regions[frame].getRegionWidth() / (float)regions[frame].getRegionHeight();
         setWidth(height * aspect);
     }
